homeworks/fastapi-basic-auth/requirements.txt
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.6.2.post1
attrs==23.2.0
Babel==2.10.3
bcc==0.29.1
bcrypt==3.2.2
blinker==1.7.0
Brlapi==0.8.5
certifi==2023.11.17
chardet==5.2.0
click==8.1.6
cloud-init==24.3.1
colorama==0.4.6
command-not-found==0.3
configobj==5.0.8
contourpy==1.3.1
cryptography==41.0.7
cupshelpers==1.0
cycler==0.12.1
dbus-python==1.3.2
defer==1.0.6
distro==1.9.0
distro-info==1.7+build1
dnspython==2.7.0
duplicity==2.1.4
email_validator==2.2.0
fastapi==0.115.5
fastapi-cli==0.0.5
fasteners==0.18
fonttools==4.55.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.20.4
httptools==0.6.4
httpx==0.27.2
idna==3.6
Jinja2==3.1.2
jsonpatch==1.32
jsonpointer==2.0
jsonschema==4.10.3
kiwisolver==1.4.7
language-selector==0.1
launchpadlib==1.11.0
lazr.restfulclient==0.14.6
lazr.uri==1.0.6
louis==3.29.0
Mako==1.3.2.dev0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.3
mdurl==0.1.2
monotonic==1.6
netaddr==0.8.0
netifaces==0.11.0
networkx==3.4.2
numpy==2.1.2
oauthlib==3.2.2
olefile==0.46
opencv-python==4.10.0.84
packaging==24.2
pandas==2.2.3
paramiko==2.12.0
passlib==1.7.4
pexpect==4.9.0
pillow==10.2.0
ptyprocess==0.7.0
pycairo==1.25.1
pycups==2.0.1
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.17.2
PyGObject==3.48.2
PyJWT==2.7.0
PyNaCl==1.5.0
pyparsing==3.1.1
PyQt5==5.15.11
PyQt5-Qt5==5.15.15
PyQt5_sip==12.15.0
pyrsistent==0.20.0
pyserial==3.5
python-apt==2.7.7+ubuntu3
python-dateutil==2.8.2
python-debian==0.1.49+ubuntu2
python-dotenv==1.0.1
python-multipart==0.0.17
pytz==2024.1
pyxdg==0.28
PyYAML==6.0.1
requests==2.31.0
rich==13.7.1
scipy==1.14.1
screen-resolution-extra==0.0.0
setuptools==68.1.2
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
starlette==0.41.3
systemd-python==235
typer==0.13.1
typing_extensions==4.10.0
tzdata==2024.2
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.2
unattended-upgrades==0.1
urllib3==2.0.7
usb-creator==0.3.16
uvicorn==0.32.0
uvloop==0.21.0
wadllib==1.3.6
watchfiles==0.24.0
websockets==14.1
wheel==0.42.0
xdg==5
xkit==0.0.0
